+++ Remaining_Shell/Win32Explorer_26.0.3.0/App_Source/RegistryAppStorageFactory.cpp
@@ -1,4 +1,4 @@
-﻿// Copyright (C) Win32Explorer Project
+// Copyright (C) Win32Explorer Project
 // SPDX-License-Identifier: GPL-3.0-only
 // See LICENSE in the top level directory
 
@@ -8,13 +8,14 @@
 #include <wil/registry.h>
 
 std::unique_ptr<RegistryAppStorage> RegistryAppStorageFactory::MaybeCreate(
-	const std::wstring &applicationKeyPath, Storage::OperationType operationType)
+	const std::wstring &applicationKeyPath, Storage::OperationType operationType, bool useHKLM)
 {
 	wil::unique_hkey applicationKey;
+	HKEY rootKey = useHKLM ? HKEY_LOCAL_MACHINE : HKEY_CURRENT_USER;
 
 	if (operationType == Storage::OperationType::Load)
 	{
-		applicationKey = OpenKeyForLoad(applicationKeyPath);
+		applicationKey = OpenKeyForLoad(applicationKeyPath, useHKLM);
 	}
 	else
 	{
@@ -22,9 +23,9 @@ std::unique_ptr<RegistryAppStorage> RegistryAppStorageFactory::MaybeCreate(
 		// important to do this to ensure that, when saving a list of items (e.g. a list of tabs or
 		// bookmarks), the existing items are removed before the updated list is stored. Otherwise,
 		// if the list shrinks, some of the previous items won't be removed.
-		SHDeleteKey(HKEY_LOCAL_MACHINE, applicationKeyPath.c_str());
+		SHDeleteKey(rootKey, applicationKeyPath.c_str());
 
-		applicationKey = CreateKeyForSave(applicationKeyPath);
+		applicationKey = CreateKeyForSave(applicationKeyPath, useHKLM);
 	}
 
 	if (!applicationKey)
@@ -35,10 +36,11 @@ std::unique_ptr<RegistryAppStorage> RegistryAppStorageFactory::MaybeCreate(
 	return std::make_unique<RegistryAppStorage>(std::move(applicationKey));
 }
 
-wil::unique_hkey RegistryAppStorageFactory::OpenKeyForLoad(const std::wstring &applicationKeyPath)
+wil::unique_hkey RegistryAppStorageFactory::OpenKeyForLoad(const std::wstring &applicationKeyPath, bool useHKLM)
 {
 	wil::unique_hkey applicationKey;
-	HRESULT hr = wil::reg::open_unique_key_nothrow(HKEY_LOCAL_MACHINE, applicationKeyPath.c_str(),
+	HKEY rootKey = useHKLM ? HKEY_LOCAL_MACHINE : HKEY_CURRENT_USER;
+	HRESULT hr = wil::reg::open_unique_key_nothrow(rootKey, applicationKeyPath.c_str(),
 		applicationKey, wil::reg::key_access::read);
 
 	if (FAILED(hr))
@@ -49,10 +51,11 @@ wil::unique_hkey RegistryAppStorageFactory::OpenKeyForLoad(const std::wstring &a
 	return applicationKey;
 }
 
-wil::unique_hkey RegistryAppStorageFactory::CreateKeyForSave(const std::wstring &applicationKeyPath)
+wil::unique_hkey RegistryAppStorageFactory::CreateKeyForSave(const std::wstring &applicationKeyPath, bool useHKLM)
 {
 	wil::unique_hkey applicationKey;
-	HRESULT hr = wil::reg::create_unique_key_nothrow(HKEY_LOCAL_MACHINE, applicationKeyPath.c_str(),
+	HKEY rootKey = useHKLM ? HKEY_LOCAL_MACHINE : HKEY_CURRENT_USER;
+	HRESULT hr = wil::reg::create_unique_key_nothrow(rootKey, applicationKeyPath.c_str(),
 		applicationKey, wil::reg::key_access::readwrite);
 
 	if (FAILED(hr))