+++ Remaining_Shell/Win32Explorer_26.0.3.0/App_Source/RegistryAppStorageFactory.h
@@ -1,4 +1,4 @@
-﻿// Copyright (C) Win32Explorer Project
+// Copyright (C) Win32Explorer Project
 // SPDX-License-Identifier: GPL-3.0-only
 // See LICENSE in the top level directory
 
@@ -13,10 +13,10 @@ class RegistryAppStorageFactory
 {
 public:
 	static std::unique_ptr<RegistryAppStorage> MaybeCreate(const std::wstring &applicationKeyPath,
-		Storage::OperationType operationType);
+		Storage::OperationType operationType, bool useHKLM = false);
 
 private:
-	static wil::unique_hkey OpenKeyForLoad(const std::wstring &applicationKeyPath);
-	static wil::unique_hkey CreateKeyForSave(const std::wstring &applicationKeyPath);
+	static wil::unique_hkey OpenKeyForLoad(const std::wstring &applicationKeyPath, bool useHKLM = false);
+	static wil::unique_hkey CreateKeyForSave(const std::wstring &applicationKeyPath, bool useHKLM = false);
 };